Corrections to Nozzle needle, and what I hope is a fix for the Actions of the new composite devices
After much debugging, the fact that there was no dispatcher object in
the Action.Target.Dispatcher call which resulted in a null object when
input binding actions were attempted on the new composite devices, was
addressed by stealing the dispatcher from the Trigger Owner.  This is
certainly not the right approach, but it seems to work and is better
that what I was attempting to get working which was a Persist Children
approach.  For now, I'm going to go with this.