new changes on BS interface, added Customgauge. Added too, Translate/rotate control by Rachmaninoff
Changes on the BS interface:
- Added ambient temperature
- Corrected UV26 power guard and Test guard buttons numbers
- Corrected System Bit selector to rotaryswitch

CustomGauge:
A new control that permits create a custom gauge with one needle. It permits add the background plate and the needle. The rotation pivot, size and position of the needle can be adjusted on the properties, so the instrument can be very customizable

Kneeboard:
- Changed the default image alignemt to Uniform
Traslate/rotate control ( by Rachmaninoff)
A new control thats permits translations in X and Y axis and rotations.
can be controlled by the cursor to move the traslations. Ideal to make horizontal or vertical levers on the cockpits, Also for create customs ADIs like the P51 one