Fix to install the 2015-2019 Visual c++ runtime
Fix for latest problem with #190 and the comments contain the design(!?!?!) thinking for the fix.  The fix files are in VC_redist.64.ClickOnce.Package.zip and need to be extracted into c:\program files (x86)\.  The specific version of vc_dist.x64.exe (14.24.28127.04) ie 2015-2019 runtime, has been loaded onto GH-Pages so that the installer can grab this specific level of the file.  This is needed because the public key is contained in the ClickOnce Bootstrapper.
I've tested this with 2017 runtime already installed and without.